# reduce allocs and map indexes for CollectPoint