… integrations/data/packets_benchmark.srpl