Adds benchmarks from @yosiat.

Two changes, instead of statsMap.Add using direct reference to *expvar.Int to avoid map look up.

Changed args to incCollected/incEmitted to not require heap allocation.

Adds benchmark for edge.CollectPoint directly. 
Improvement is good:

```
benchmark                    old ns/op     new ns/op     delta
Benchmark_CollectPoint-8     1031          707           -31.43%

benchmark                    old allocs     new allocs     delta
Benchmark_CollectPoint-8     2              0              -100.00%

benchmark                    old bytes     new bytes     delta
Benchmark_CollectPoint-8     256           0             -100.00%
```

Improvement on full write test is good too.
```
benchmark                                            old ns/op       new ns/op      delta
Benchmark_Write_MeasurementNameNotMatches_1000-4     6273967891      4240956941     -32.40%
Benchmark_Write_MeasurementNameMatches_1000-4        6708792279      4503474984     -32.87%
Benchmark_Write_MeasurementNameNotMatches_100-4      7768092706      4973414305     -35.98%
Benchmark_Write_MeasurementNameMatches_100-4         8383660564      5157680889     -38.48%
Benchmark_Write_MeasurementNameNotMatches_10-4       9831826369      5263307437     -46.47%
Benchmark_Write_MeasurementNameMatches_10-4          10232057586     5258015149     -48.61%

benchmark                                            old allocs     new allocs     delta
Benchmark_Write_MeasurementNameNotMatches_1000-4     15462870       57439          -99.63%
Benchmark_Write_MeasurementNameMatches_1000-4        19313940       3901498        -79.80%
Benchmark_Write_MeasurementNameNotMatches_100-4      23090665       4292295        -81.41%
Benchmark_Write_MeasurementNameMatches_100-4         23742484       4800160        -79.78%
Benchmark_Write_MeasurementNameNotMatches_10-4       24095167       4775145        -80.18%
Benchmark_Write_MeasurementNameMatches_10-4          24147784       4890630        -79.75%

benchmark                                            old bytes      new bytes     delta
Benchmark_Write_MeasurementNameNotMatches_1000-4     1975214888     4262320       -99.78%
Benchmark_Write_MeasurementNameMatches_1000-4        2405864264     434731368     -81.93%
Benchmark_Write_MeasurementNameNotMatches_100-4      2882851768     478502496     -83.40%
Benchmark_Write_MeasurementNameMatches_100-4         2957967752     535381800     -81.90%
Benchmark_Write_MeasurementNameNotMatches_10-4       3002221608     532581984     -82.26%
Benchmark_Write_MeasurementNameMatches_10-4          3008055144     545516640     -81.86%
```

CollectPoint is still spending lots of time counting stats, I am playing around with channels based implementation vs atmoic.Add