@nathanielc 2016-03-18
Using a channel based implementation instead of atomic.Add was not faster.