@yosiat 2016-03-18
@nathanielc I think there is a bug in the Benchmark, I am not cleaning after each benchmark the TaskMaster, so we are actually not testing against 100 tasks or 10 tasks.

I am fixing this and will upload new version to yosiat/kapacitor/perf