@nathanielc 2016-03-18
@yosiat Ok thanks, either way the CollectPoint benchmark is -30% on its own which seems to translating well to the broader test.