@yosiat 2016-03-18
Fixed this:
```
Benchmark_Write_MeasurementNameNotMatches_1000-4	       1	11760042929 ns/op	1975218280 B/op	15468357 allocs/op
Benchmark_Write_MeasurementNameMatches_1000-4   	       1	11509590974 ns/op	2406124968 B/op	19315430 allocs/op
Benchmark_Write_MeasurementNameNotMatches_100-4 	    5000	    205519 ns/op	   54418 B/op	     448 allocs/op
Benchmark_Write_MeasurementNameMatches_100-4    	    5000	    205267 ns/op	   65278 B/op	     545 allocs/op
Benchmark_Write_MeasurementNameNotMatches_10-4  	   50000	     24441 ns/op	    4359 B/op	      53 allocs/op
Benchmark_Write_MeasurementNameMatches_10-4     	   50000	     26836 ns/op	    4468 B/op	      54 allocs/op
ok  	command-line-arguments	84.642s
```

https://github.com/yosiat/kapacitor/commit/22d4616d10d8b90ecfaf059da624e4fc63a4535d