@yosiat 2016-03-18
@nathanielc I think a possible performance improvement is to do some api breaking,
and changed "StreamCollect" to accept an array of points.

Therefore, you can increment once for multiple points.