@nathanielc 2016-03-18
Here are the correct results, basically the same...
```
benchmark                                            old ns/op      new ns/op      delta
Benchmark_Write_MeasurementNameNotMatches_1000-4     7307662678     4678726431     -35.98%
Benchmark_Write_MeasurementNameMatches_1000-4        6724460416     4460785463     -33.66%
Benchmark_Write_MeasurementNameNotMatches_100-4      23275          15479          -33.50%
Benchmark_Write_MeasurementNameMatches_100-4         23620          15570          -34.08%
Benchmark_Write_MeasurementNameNotMatches_10-4       15458          14639          -5.30%
Benchmark_Write_MeasurementNameMatches_10-4          15396          14486          -5.91%

benchmark                                            old allocs     new allocs     delta
Benchmark_Write_MeasurementNameNotMatches_1000-4     15466521       57447          -99.63%
Benchmark_Write_MeasurementNameMatches_1000-4        19309921       3904833        -79.78%
Benchmark_Write_MeasurementNameNotMatches_100-4      88             48             -45.45%
Benchmark_Write_MeasurementNameMatches_100-4         97             53             -45.36%
Benchmark_Write_MeasurementNameNotMatches_10-4       50             48             -4.00%
Benchmark_Write_MeasurementNameMatches_10-4          51             49             -3.92%

benchmark                                            old bytes      new bytes     delta
Benchmark_Write_MeasurementNameNotMatches_1000-4     1975554248     4264496       -99.78%
Benchmark_Write_MeasurementNameMatches_1000-4        2406015896     435170392     -81.91%
Benchmark_Write_MeasurementNameNotMatches_100-4      8872           3841          -56.71%
Benchmark_Write_MeasurementNameMatches_100-4         9958           4384          -55.98%
Benchmark_Write_MeasurementNameNotMatches_10-4       4104           3843          -6.36%
Benchmark_Write_MeasurementNameMatches_10-4          4157           3896          -6.28%
```