@nathanielc 2016-03-18
@yosiat Breaking the API is not a possibility, at least not right now. 

The change you are proposing already exists as CollectBatch. A batch is an array of points with some consolidated metadata.  CollectPoint exists because it was needed in addition to CollectBatch. 

That being said, it doesn't mean we can't group updates and only perform `.Add` every so often. I am playing with something like that now.