@nathanielc 2016-03-18
Another significant improvement if groupBy is not being used:
```
benchmark                                            old ns/op      new ns/op      delta
Benchmark_Write_MeasurementNameNotMatches_1000-4     5159738246     3880726492     -24.79%
Benchmark_Write_MeasurementNameMatches_1000-4        4020222956     3481916661     -13.39%
Benchmark_Write_MeasurementNameNotMatches_100-4      15369          15314          -0.36%
Benchmark_Write_MeasurementNameMatches_100-4         15643          15226          -2.67%
Benchmark_Write_MeasurementNameNotMatches_10-4       14435          13266          -8.10%
Benchmark_Write_MeasurementNameMatches_10-4          14664          13217          -9.87%

benchmark                                            old allocs     new allocs     delta
Benchmark_Write_MeasurementNameNotMatches_1000-4     57444          57440          -0.01%
Benchmark_Write_MeasurementNameMatches_1000-4        3905954        57426          -98.53%
Benchmark_Write_MeasurementNameNotMatches_100-4      48             48             +0.00%
Benchmark_Write_MeasurementNameMatches_100-4         53             48             -9.43%
Benchmark_Write_MeasurementNameNotMatches_10-4       48             48             +0.00%
Benchmark_Write_MeasurementNameMatches_10-4          49             48             -2.04%

benchmark                                            old bytes     new bytes     delta
Benchmark_Write_MeasurementNameNotMatches_1000-4     4262656       4262384       -0.01%
Benchmark_Write_MeasurementNameMatches_1000-4        435289880     4261568       -99.02%
Benchmark_Write_MeasurementNameNotMatches_100-4      3841          3842          +0.03%
Benchmark_Write_MeasurementNameMatches_100-4         4385          3842          -12.38%
Benchmark_Write_MeasurementNameNotMatches_10-4       3842          3841          -0.03%
Benchmark_Write_MeasurementNameMatches_10-4          3897          3841          -1.44%
```