@yosiat 2016-03-18
@nathanielc why you didn't used "buffer.WriteString" method?