@nathanielc 2016-03-18
@yosiat I just missed it, it won't make a performance difference. I'll update it soon